requirements.txt
matplotlib
pandas
numpy
plotly
streamlit_plotly_events